Exceptions, namedQuery and throwing exceptions

+++ CarAppBackend/src/main/java/backend/service/impl/CarServiceImpl.java
@@ -8,20 +8,20 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
 import backend.entity.CarEntity;
+import backend.exception.CarFoundException;
+import backend.exception.CarNotFoundException;
 import backend.service.CarService;
 
-@Stateless(name="carService")
+@Stateless
 public class CarServiceImpl implements CarService {
 
-	private static final String findAllCarsQuery = "SELECT c FROM CarEntity c";
-	
 	@PersistenceContext(unitName="carEntityPU")
 	private EntityManager em;
 	
 	@SuppressWarnings("unchecked")
 	@Override
 	public List<CarEntity> getAllCars() {
-		return (List<CarEntity>) em.createQuery(findAllCarsQuery).getResultList();
+		return (List<CarEntity>) em.createNamedQuery("CarService.findAllCars").getResultList();
 	}
 
 	@Override
@@ -30,25 +30,32 @@ public class CarServiceImpl implements CarService {
 	}
 
 	@Override
-	public CarEntity createCar(CarEntity car) {
+	public CarEntity createCar(CarEntity car) throws CarFoundException {
 		try {
 			em.persist(car);
 			return car;
 		} catch (EntityExistsException ex) {
-			return null;
+			throw new CarFoundException(car);
 		}
 	}
 
 	@Override
-	public CarEntity updateCar(CarEntity car) {
+	public CarEntity updateCar(CarEntity car) throws CarNotFoundException {
+		CarEntity auxCar = em.find(CarEntity.class, car.getId());
+		if(auxCar == null) {
+			throw new CarNotFoundException(car.getId());
+		}
 		return em.merge(car);
 	}
 
 	@Override
-	public void deleteCar(String id) {
-		CarEntity car = new CarEntity();
-		car.setId(id);
+	public boolean deleteCar(String id) throws CarNotFoundException {
+		CarEntity car = em.find(CarEntity.class, id);
+		if(car == null) {
+			throw new CarNotFoundException(id);
+		}
 		em.remove(car);
+		return true;
 	}
 
 }
